fixed department issue

+++ src/main/java/com/tumbwe/examandclassattendanceapi/dto/DepartmentDto.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Setter
 public class DepartmentDto {
     private String name;
-    private int schoolId;
+    private Long schoolId;
     private Long id;
 }
